finished syncing of contracts with all sub types, but not deleting bugdets yet

+++ mssql-database/BoatExt_PriceCategories.sql
@@ -2,6 +2,6 @@
 (
 	[Id] INT NOT NULL PRIMARY KEY, 
     [Name] NVARCHAR(50) NOT NULL, 
-    [PricePerUnit] INT NOT NULL,
-    [MinutesPerDay] INT NOT NULL,
+    [PricePerUnit] FLOAT NOT NULL,
+    [MinutesPerDay] FLOAT NOT NULL,
 )
